requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.63.0
docling-core==2.54.0
docling-ibm-models==3.10.2
docling-parse==4.7.1
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==38.2.0
fastapi==0.123.0
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
gradio==6.0.1
gradio_client==2.0.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.1.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-docling==2.0.0
langchain-google-genai==3.2.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.1.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.4.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
pypdf==6.4.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
rapidocr==3.4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.5.1
typer==0.19.2
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
